3_scanner/requirements.txt
fastapi==0.128.0
uvicorn==0.30.0
pydantic==2.12.5
python-dotenv==1.2.1
GitPython==3.1.40
PyYAML==6.0.3
requests==2.32.3
packaging==24.0

pytest==8.3.4
